+++ tests/citadel.rs
@@ -10,8 +10,9 @@ use dusk_plonk::prelude::*;
 
 static LABEL: &[u8; 12] = b"dusk-network";
 
-const CAPACITY: usize = 15; // capacity required for the setup
-const DEPTH: usize = 17; // depth of the 4-ary Merkle tree
+const CAPACITY: usize = 16; // capacity required for the setup
+const DEPTH: usize = 17; // depth of the n-ary Merkle tree
+pub const ARITY: usize = 4; // arity of the Merkle tree
 
 use zk_citadel::gadget;
 use zk_citadel::license::{License, LicenseProverParameters, Request, Session, SessionCookie};
@@ -21,12 +22,12 @@ use rand_core::{CryptoRng, OsRng, RngCore};
 
 #[derive(Default, Debug)]
 pub struct Citadel {
-    lpp: LicenseProverParameters<DEPTH>,
+    lpp: LicenseProverParameters<DEPTH, ARITY>,
     sc: SessionCookie,
 }
 
 impl Citadel {
-    pub fn new(lpp: &LicenseProverParameters<DEPTH>, sc: &SessionCookie) -> Self {
+    pub fn new(lpp: &LicenseProverParameters<DEPTH, ARITY>, sc: &SessionCookie) -> Self {
         Self { lpp: *lpp, sc: *sc }
     }
 }
@@ -43,7 +44,11 @@ impl Circuit for Citadel {
 
 fn compute_random_license<R: RngCore + CryptoRng>(
     rng: &mut R,
-) -> (License, LicenseProverParameters<DEPTH>, SessionCookie) {
+) -> (
+    License,
+    LicenseProverParameters<DEPTH, ARITY>,
+    SessionCookie,
+) {
     // Example values
     const USER_ATTRIBUTES: u64 = 112233445566778899u64;
     const CHALLENGE: u64 = 20221126u64;
@@ -63,9 +68,9 @@ fn compute_random_license<R: RngCore + CryptoRng>(
 
     // Second, the LP computes these values and grants the License
     let attr = JubJubScalar::from(USER_ATTRIBUTES);
-    let lic = License::new(&attr, &ssk_lp, &req, rng);
+    let mut lic = License::new(&attr, &ssk_lp, &req, rng);
     let mut state = State::new(); // the compiler takes DEPTH from expected 'lpp' to return
-    state.append_license(&lic);
+    state.append_license(&mut lic);
 
     // Third, the user computes these values to generate the ZKP later on
     let vk = ssk.view_key();